pyproject.toml
[project]
name = "gps-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.104.1",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "requests>=2.32.5",
    "uvicorn[standard]==0.24.0",
    "websockets==12.0",
]
